First commit of entire workspace
This commit represents the culmination of a large amount of work, which should've been broken up
into separate commits. Alas, the work was largely experimental until quite a few of these parts had
solidified, so this is what we get.

General layout: functionally equivalent services are put into their own namespace, while shared
services are put into the `default` namespace. Each service is accompanied by a `Dockerfile` and a
`service.yaml` file, which deploys both the `Service` object and any `Deployment` objects to the
relevant namespace. Services that require secrets or config will generally have those as separate
template files in the same directory.

The `Makefile` in the root directory orchestrates or simplifies deployments. Of course, the right
incantation will be sufficient for deploying the Docker images and K8s services. Template files,
however, rely on the Makefile to perform a series of esoteric processes before applying to K8s, and
therefore are considered as non-standard configuration files.

There is quite a bit of work remaining, including:

  - Setting up TLS certificates for Voyager via Let's Encrypt.
  - Instrumentation and logging.
  - Service deployment via internal Git hooks, if possible.

Plus other bits I may have forgotten about.