Fix nginx, php-fpm Docker images and deployments
Both `nginx` and `php-fpm` Docker containers are meant to be colocated with the application, and
it's much easier to do so when running inside the same Pod with a shared volume mount.

The `nginx` Docker image has been moved to Debian Stretch, to mirror the `php-fpm` image, and some
issues with logging have been fixed.